Cast partitionNum to Int (#91)

+++ nebula-algorithm/src/test/scala/com/vesoft/nebula/algorithm/config/ConfigSuite.scala
@@ -46,7 +46,7 @@ class ConfigSuite {
     assert(sparkConfig.map.size == 3)
 
     val spark = SparkConfig.getSpark(configs)
-    assert(spark.partitionNum.toInt == 100)
+    assert(spark.partitionNum == 100)
   }
 
   @Test
